# apps/reference/specs/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: Human-Agent API Platform
  version: '2025-01-29'
  description: |
    API platform designed as a first-class experience for both human developers and AI agents.

    **Key Features:**
    - Schema-first design with complete OpenAPI specs
    - Structured error responses with actionable suggestions
    - Agent-aware rate limiting (100 human / 500 agent requests per minute)
    - Header-based versioning (API-Version: YYYY-MM-DD)
    - Dry-run mode for all mutations (?dry_run=true)
    - Automatic agent detection and tracking

    **For AI Agents:**
    - All parameters include descriptions and examples
    - Errors include actionable suggestions for self-correction
    - Rate limit headers enable proactive throttling
    - Retry-After headers prevent retry storms

  contact:
    name: API Platform Team
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development
  - url: https://api.example.com
    description: Production
...
paths:
...
  /api/approvals:
    get:
      operationId: listApprovals
      summary: List pending changes awaiting approval
      tags:
        - approvals
      security:
        - bearerAuth: []
      description: Lists pending changes owned by the authenticated user (admins see all).
      responses:
        '200':
          description: Approvals visible to the caller
          content:
            application/json:
              schema:
                type: object
                description: Approval list
                properties:
                  data:
                    type: object
                    description: Wrapper object
                    properties:
                      approvals:
                        type: array
                        description: The pending changes
                        items:
                          $ref: '#/components/schemas/PendingChange'
              example:
                data:
                  approvals: []
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/approvals/{id}:
    get:
      operationId: getApproval
      summary: Approval status
      tags:
        - approvals
      security:
        - bearerAuth: []
      description: Returns the pending change. Visible to the owning user, the proposing agent, and admins.
      parameters:
        - name: id
          in: path
          required: true
          description: Approval id
          example: approval_1
          schema:
            type: string
      responses:
        '200':
          description: The pending change
          content:
            application/json:
              schema:
                type: object
                description: Approval wrapper
                properties:
                  data:
                    $ref: '#/components/schemas/PendingChange'
              example:
                data:
                  id: approval_1
                  status: pending
                  summary: DELETE /api/v2/tasks/task_1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/approvals/{id}/events:
    get:
      operationId: streamApprovalEvents
      summary: Stream approval resolution (SSE)
      tags:
        - approvals
      security:
        - bearerAuth: []
      description: |
        Server-sent events stream. Emits the current status immediately,
        then a terminal event when the approval is resolved — agents learn
        the outcome without polling. The stream closes after resolution.
      parameters:
        - name: id
          in: path
          required: true
          description: Approval id
          example: approval_1
          schema:
            type: string
      responses:
        '200':
          description: SSE stream of approval events
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE frames with JSON payloads ({approval_id, status, result_status})
              example: 'data: {"approval_id":"approval_1","status":"executed","result_status":201}'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/approvals/{id}/approve:
    post:
      operationId: approvePendingChange
      summary: Approve and execute a pending change
      tags:
        - approvals
      security:
        - bearerAuth: []
      description: |
        Human consent (session token required). The captured request is
        executed through the full middleware stack under a single-use
        execution token restoring the proposer's context; the result is
        stored on the approval and streamed to SSE subscribers.
      parameters:
        - name: id
          in: path
          required: true
          description: Approval id
          example: approval_1
          schema:
            type: string
      responses:
        '200':
          description: Executed (status executed/failed with the result)
          content:
            application/json:
              schema:
                type: object
                description: Resolved approval
                properties:
                  data:
                    $ref: '#/components/schemas/PendingChange'
              example:
                data:
                  id: approval_1
                  status: executed
                  result_status: 204
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/approvals/{id}/reject:
    post:
      operationId: rejectPendingChange
      summary: Reject a pending change
      tags:
        - approvals
      security:
        - bearerAuth: []
      description: Human consent withheld (session token required). The change never executes.
      parameters:
        - name: id
          in: path
          required: true
          description: Approval id
          example: approval_1
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              description: Optional rejection details
              properties:
                reason:
                  type: string
                  description: Why the change was rejected (shown to the proposer)
            example:
              reason: Not needed anymore
      responses:
        '200':
          description: Rejected
          content:
            application/json:
              schema:
                type: object
                description: Resolved approval
                properties:
                  data:
                    $ref: '#/components/schemas/PendingChange'
              example:
                data:
                  id: approval_1
                  status: rejected
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'

# ==============================================================================
# Components
# ==============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
...
  headers:
    X-Request-ID:
      description: Unique request identifier for tracing and support
      schema:
        type: string
        example: req_abc123
...
    X-RateLimit-Limit:
      description: Maximum requests allowed in the current window
      schema:
        type: integer
        example: 500
    X-RateLimit-Remaining:
      description: Requests remaining in the current window
      schema:
        type: integer
        example: 387
...
    Retry-After:
      description: Seconds to wait before retrying (set when rate limited)
      schema:
        type: integer
        example: 42
  schemas:
...
    PendingChange:
      type: object
      description: A captured mutation awaiting (or resolved by) human approval
      properties:
        id:
          type: string
          description: Approval id
          example: approval_1
        status:
          type: string
          description: pending, approved, rejected, executed, failed, or expired
          enum: [pending, approved, rejected, executed, failed, expired]
          example: pending
        summary:
          type: string
          description: Human-readable summary of the captured request
          example: DELETE /api/v2/tasks/task_1
        method:
          type: string
          description: Captured HTTP method
          example: DELETE
        path:
          type: string
          description: Captured request path
          example: /api/v2/tasks/task_1
        proposer_agent_id:
          type: string
          description: Agent that proposed the change (when delegated)
          example: agent_1
        proposer_token_use:
          type: string
          description: session or delegated
          example: delegated
        owner_user_id:
          type: string
          description: User whose consent resolves this change
          example: user_42
        result_status:
          type: integer
          description: HTTP status of the executed request (after approval)
          example: 204
        reject_reason:
          type: string
          description: Reason given on rejection
        expires_at:
          type: string
          format: date-time
          description: Pending changes expire if unresolved
        created_at:
          type: string
          format: date-time
          description: Proposal time
        resolved_at:
          type: string
          format: date-time
          description: Resolution time
        resolved_by:
          type: string
          description: User who resolved it
          example: user_42

    # Delegation schemas
...
    ErrorResponse:
      type: object
      description: Standard error response envelope containing error details
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      description: Detailed error information following RFC 7807 problem details pattern
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_PARAMETER
            - MISSING_REQUIRED_FIELD
            - VALUE_OUT_OF_RANGE
            - INVALID_FORMAT
            - RESOURCE_NOT_FOUND
            - CONFLICT
            - UNAUTHORIZED
            - FORBIDDEN
            - RATE_LIMIT_EXCEEDED
            - INTERNAL_SERVER_ERROR
            - SERVICE_UNAVAILABLE
          example: INVALID_PARAMETER
        message:
          type: string
          description: Human-readable error description
          example: The limit parameter is out of range
        target:
          type: string
          description: The field or parameter that caused the error
          example: limit
        details:
          type: array
          description: Additional error details with actionable suggestions
          items:
            $ref: '#/components/schemas/ErrorDetail'
        doc_url:
          type: string
          format: uri
          description: URL to error documentation
          example: https://docs.example.com/errors/INVALID_PARAMETER
        request_id:
          type: string
          description: Unique request identifier for support
          example: req_abc123
    ErrorDetail:
      type: object
      description: Granular error detail with actionable suggestions for agent self-correction
      required:
        - code
        - message
        - suggestion
      properties:
        code:
          type: string
          description: Specific error detail code
          example: VALUE_OUT_OF_RANGE
        message:
          type: string
          description: Detailed error message
          example: limit must be between 1 and 100
        suggestion:
          type: string
          description: Actionable suggestion for fixing the error (REQUIRED for agent self-correction)
          example: Set limit to a value between 1 and 100, or omit it to use the default of 20
        target:
          type: string
          description: Specific field this detail relates to
          example: limit
  responses:
    BadRequest:
      description: Invalid request parameters
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INVALID_PARAMETER
              message: The limit parameter is out of range
              target: limit
              details:
                - code: VALUE_OUT_OF_RANGE
                  message: limit must be between 1 and 100
                  suggestion: Set limit to a value between 1 and 100, or omit it to use the default of 20
                  target: limit
              request_id: req_abc123
    Unauthorized:
      description: Authentication required
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
              details:
                - code: MISSING_TOKEN
                  message: No Bearer token provided
                  suggestion: Include Authorization header with a valid Bearer token from /api/auth/login
              request_id: req_abc123
...
    NotFound:
      description: Resource not found
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: The requested resource was not found
              target: id
              details:
                - code: NOT_FOUND
                  message: The requested resource does not exist
                  suggestion: Check the ID and try again, or list resources to find the correct ID
                  target: id
              request_id: req_abc123
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
        Retry-After:
          $ref: '#/components/headers/Retry-After'
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Rate limit exceeded
              details:
                - code: TOO_MANY_REQUESTS
                  message: You have exceeded the rate limit of 500 requests per 60 seconds
                  suggestion: Wait 42 seconds before retrying. Implement exponential backoff to avoid retry storms.
                  target: rate_limit
              request_id: req_abc123